variable length now seems to match performance between rollout and train, needs code cleanup

+++ ltron_torch/gym_tensor.py
@@ -23,7 +23,8 @@ tensor_mean = torch.FloatTensor(default_mean)
 tensor_std = torch.FloatTensor(default_std)
 
 default_tile_transform = transforms.Compose([
-    lambda x : torch.FloatTensor(x)/255.,
+    #lambda x : torch.FloatTensor(x)/255.,
+    lambda x : x / 255.,
     lambda x : (x - tensor_mean.view(1,1,1,3)) / tensor_std.view(1,1,1,3)
 ])
